move nat/ changes to dragent/
